requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
fastapi==0.115.7
fastapi-cli==0.0.7
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.11.14
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gotrue==2.11.2
greenlet==3.1.1
groq==0.15.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx
httpx-sse==0.4.0
huggingface-hub==0.28.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.49
langchain-google-community==2.0.7
langchain-google-vertexai
langchain-groq==0.2.3
langchain-neo4j==0.2.0
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-pinecone==0.2.2
langchain-postgres==0.0.12
langchain-tests==0.3.10
langchain-text-splitters==0.3.5
langchain-together==0.3.0
langgraph==0.2.63
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.11
llmsherpa==0.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.22.0
neo4j==5.27.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.6
openai==1.59.7
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pgvector==0.2.5
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
postgrest==0.19.3
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
PyPDF2==3.0.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
realtime==2.0.6
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
storage3==0.11.1
streamlit==1.41.1
streamlit-authenticator==0.4.1
streamlit-google-auth==1.1.8
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.2
sympy==1.13.1
syrupy==4.8.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
vertexai
watchdog==6.0.0
watchfiles==1.0.4
websockets==12.0
yarl==1.18.3
